Michael,Franz, and I Worked on the logic for different views on Carousel- WIP
