+++ Web/src/actions/constants/auth.js
@@ -1,0 +1,4 @@
+export const CHECK_IF_AUTHENTICATED = 'CHECK_IF_AUTHENTICATED';
+export const USER_UNAUTHENTICATED = 'USER_UNAUTHENTICATED';
+export const USER_AUTHENTICATED = 'USER_AUTHENTICATED';
+export const AUTHENTICATION_ERROR = 'AUTHENTICATION_ERROR';
