+++ Web/src/actions/constants/carousel.js
@@ -1,0 +1,7 @@
+export const CAROUSEL_ITEMS_ERROR = "CAROUSEL_ITEMS_ERRORS";
+export const CAROUSEL_ITEMS_REGISTERED = "CAROUSEL_ITEMS_REGISTERED";
+export const CAROUSEL_ITEMS_UNREGISTERED = "CAROUSEL_ITEMS_UNREGISTERED";
+export const CAROUSEL_ITEMS_DEFAULT = "CAROUSEL_ITEMS_DEFAULT";
+export const CAROUSEL_ITEMS_NOT_REGISTERED ="CAROUSEL_ITEMS_NOT_REGISTERED"
+
+
