+++ Web/src/actions/constants/pictures.js
@@ -1,0 +1,2 @@
+export const FETCH_PICTURE_ERROR = "FETCH_PICTURE_ERROR";
+export const FETCH_PICTURES = "FETCH_PICTURES";
