+++ Web/src/actions/constants/register.js
@@ -1,0 +1,5 @@
+export const CHECK_IF_REGISTERED = 'CHECK_IF_REGISTERED';
+export const USER_REGISTERED = 'USER_REGISTERED';
+export const REGISTRATION_ERROR = 'REGISTRATION_ERROR';
+export const REGISTRATION_EXPIRED = 'REGISTRATION_EXPIRED';
+export const SIGN_UP_TOGGLE = 'SIGN_UP_TOGGLE';
