+++ Web/src/components/Carousel.js
@@ -1,36 +1,48 @@
-import React from 'react';
-import { UncontrolledCarousel } from 'reactstrap';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { withRouter, Link } from 'react-router-dom';
+import { Button, UncontrolledCarousel } from 'reactstrap';
+import {signUpToggle, getCarouselItems} from '../actions';
 
-// TODO:
-// right now this is just boiler plate stuff
-// it all will need updated
-// Update const items to fit what we would like 
-// shown on the first page
-// Probably a demo of what our service does
-// we'll need pictures
-// we most likely will want to demo how to get started usign our service
 
-const items = [
-  {
-    src: 'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_15ba800aa1d%20text%20%7B%20fill%3A%23555%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_15ba800aa1d%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23777%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22285.921875%22%20y%3D%22218.3%22%3EFirst%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
-    altText: 'Slide 1',
-    caption: 'Slide 1',
-    header: 'Slide 1 Header'
-  },
-  {
-    src: 'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_15ba800aa20%20text%20%7B%20fill%3A%23444%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_15ba800aa20%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23666%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22247.3203125%22%20y%3D%22218.3%22%3ESecond%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
-    altText: 'Slide 2',
-    caption: 'Slide 2',
-    header: 'Slide 2 Header'
-  },
-  {
-    src: 'data:image/svg+xml;charset=UTF-8,%3Csvg%20width%3D%22800%22%20height%3D%22400%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%20800%20400%22%20preserveAspectRatio%3D%22none%22%3E%3Cdefs%3E%3Cstyle%20type%3D%22text%2Fcss%22%3E%23holder_15ba800aa21%20text%20%7B%20fill%3A%23333%3Bfont-weight%3Anormal%3Bfont-family%3AHelvetica%2C%20monospace%3Bfont-size%3A40pt%20%7D%20%3C%2Fstyle%3E%3C%2Fdefs%3E%3Cg%20id%3D%22holder_15ba800aa21%22%3E%3Crect%20width%3D%22800%22%20height%3D%22400%22%20fill%3D%22%23555%22%3E%3C%2Frect%3E%3Cg%3E%3Ctext%20x%3D%22277%22%20y%3D%22218.3%22%3EThird%20slide%3C%2Ftext%3E%3C%2Fg%3E%3C%2Fg%3E%3C%2Fsvg%3E',
-    altText: 'Slide 3',
-    caption: 'Slide 3',
-    header: 'Slide 3 Header'
-  }
-];
 
-const Carousel = (props) => <UncontrolledCarousel items={items} />;
+class DefaultCarousel extends Component {
+    componentDidMount() {
+      this.props.getCarouselItems(this.props.auth);
+    }
 
-export default Carousel;
+    toggle() {
+        this.props.signUpToggle(this.props.modal)
+    }
+    
+    renderCarouselGroup(items) {
+        return ( <UncontrolledCarousel items={items} />)
+        
+    }
+    
+    getItems() {
+        console.log(this)
+        const {registered, authenticated} = this.props.auth;
+        if(!registered && !authenticated) { 
+           return this.renderCarouselGroup(this.props.carouselItems);
+        } else if(!registered && authenticated) {
+            return this.renderCarouselGroup(this.props.carouselItems);
+        } else {
+            return this.renderCarouselGroup(this.props.carouselItems);
+        }
+    }
+    
+    render() {
+      return(this.getItems.bind(this))
+    }
+}
+
+const mapStateToProps = state => {
+    return {
+      auth: state.auth,
+      modal: state.modal,
+      carouselItems: state.carouselItems 
+    };
+};
+
+export default withRouter(connect(mapStateToProps, {signUpToggle, getCarouselItems})(DefaultCarousel));
