+++ Web/src/components/DefaultPage.js
@@ -1,31 +1,11 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { withRouter, Link } from 'react-router-dom';
-import {Button} from 'reactstrap';
+import React from 'react';
 import Header from './Header';
-import Carousel from './Carousel';
-import {signUpToggle} from '../actions'
-class DefaultPage extends Component {
-  toggle(){
-    this.props.signUpToggle(this.props.modal)
-  }
-  render() {
-    return (
-      <div>
+import DefaultCarousel from './Carousel';
+import { Container } from 'reactstrap';
+
+export default props =>
+      <Container fluid>
           <Header/>
-          <Carousel/>
-          <h1> LINKSTAGRAM </h1>
-          {!this.props.registered ? <Button onClick={this.toggle.bind(this)}>BUY NOW</Button> : ''}
-      </div>
-      ); 
-    }
-}
-
-const mapStateToProps = state => {
-  return {
-    auth: state.auth,
-    modal: state.modal
-  };
-};
-
-export default withRouter(connect(mapStateToProps, {signUpToggle})(DefaultPage));
+          <DefaultCarousel/>
+      </Container>
+
