+++ Web/src/components/Modals/SignUpModal.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import { connect } from 'react-redux';
 import { withRouter, Link } from 'react-router-dom';
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
-import {signup,signUpToggle} from '../../actions';
+import {signup, signUpToggle} from '../../actions';
 
 class SignUpModal extends Component { 
     constructor(props){
@@ -33,4 +33,4 @@ const mapStateToProps = state => {
     };
 };
   
-export default withRouter(connect(mapStateToProps, {signUpToggle})(SignUpModal));
+export default withRouter(connect(mapStateToProps)(SignUpModal));
