+++ Web/src/reducers/billing.js
@@ -7,4 +7,4 @@ export default (state = false, action) => {
       default:
         return state;
     }
-}
+}
