+++ Web/src/reducers/index.js
@@ -2,12 +2,14 @@ import { combineReducers } from 'redux';
 import AuthReducer from './auth';
 import { reducer as FormReducer } from 'redux-form';
 import  BillingReducer from './billing';
-import fetchImagesReducer from './fetchImages';
+import FetchImagesReducer from './fetchImages';
+import CarouselReducer from './carousels';
 
 const rootReducer = combineReducers({
   auth: AuthReducer,
   form: FormReducer,
   modal: BillingReducer,
-  pictures: fetchImagesReducer
+  pictures: FetchImagesReducer,
+  carouselItems: CarouselReducer
 });
 export default rootReducer;
